rest_album_of_songs/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==22.1.0
blue==0.9.1
click==8.1.7
Django==4.1.5
djangorestframework==3.14.0
drf-spectacular==0.26.4
flake8==4.0.1
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
mccabe==0.6.1
mypy-extensions==1.0.0
pathspec==0.11.2
platformdirs==3.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
sqlparse==0.4.4
tomli==2.0.1
uritemplate==4.1.1
